+++ sel_compiler/src/tests/touch.rs
@@ -65,7 +65,7 @@ fn compiles_touch_integer() {
     let tree = compiler.compile(&input);
 
     let root = tree.get_root();
-    let root_value: i32 = tree.get_integer_value_of(&root).unwrap();
+    let root_value: i64 = tree.get_integer_value_of(&root).unwrap();
 
     assert_eq!(root.get_operation(), Operation::Touch);
     assert_eq!(root.get_data_type(), DataType::Integer);
